+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+### Changed
+- You must now pass a positive integer for `maxAge` (instead of any positive number)
+- You cannot pass `undefined` for `maxAge` (though you can still omit the property)
+
 ## 0.2.0 - 2019-05-04
 ### Added
 - TypeScript type definitions. See [helmetjs/helmet#188](https://github.com/helmetjs/helmet/issues/188)